Move the code under Tools (and the tests under test_tools).
