+++ Lib/test/test_tools/test_c_analyzer/test_c_statics/test_info.py
@@ -1,7 +1,9 @@
 import string
 import unittest
 
-from test.test_c_statics.cg.info import (
+from .. import tool_imports_for_tests
+with tool_imports_for_tests():
+    from c_statics.info import (
         normalize_vartype, Symbol, StaticVar,
         )
