+++ Lib/test/test_tools/test_c_analyzer/test_c_statics/test_parse.py
@@ -1,7 +1,9 @@
 import textwrap
 import unittest
 
-from test.test_c_statics.cg.parse import (
+from .. import tool_imports_for_tests
+with tool_imports_for_tests():
+    from c_statics.parse import (
         iter_global_declarations, iter_local_statements,
         parse_func, parse_var, parse_compound,
         iter_variables,
