+++ Lib/test/test_tools/test_c_analyzer/test_c_statics/test_supported.py
@@ -1,7 +1,9 @@
 import unittest
 
-from test.test_c_statics.cg import info
-from test.test_c_statics.cg.supported import is_supported
+from .. import tool_imports_for_tests
+with tool_imports_for_tests():
+    from c_statics import info
+    from c_statics.supported import is_supported
 
 
 class IsSupportedTests(unittest.TestCase):
