+++ Tools/c-analyzer/c-statics.py
