+++ Tools/c-analyzer/c_statics/__init__.py
