+++ Tools/c-analyzer/c_statics/__main__.py
