+++ Tools/c-analyzer/c_statics/info.py
