+++ Tools/c-analyzer/c_statics/parse.py
