+++ Tools/c-analyzer/c_statics/scan.py
