+++ Tools/c-analyzer/c_statics/show.py
