+++ Tools/c-analyzer/c_statics/symbols.py
